refactor(build): split the page generator into bin/pages/
1476 lines in one file, of which 29% was data that has nothing to do with
rendering: 220 lines of per-locale chrome strings and 207 lines of page
inventory. Now:

  bin/build-pages.mjs   132  assembly + CLI, re-exporting the public surface
  bin/pages/render-home  358  the homepage template
  bin/pages/render-page  239  the template for everything else
  bin/pages/ui           227  per-locale chrome strings
  bin/pages/pages        212  the page inventory
  bin/pages/markdown     127  frontmatter, renderer, extractors
  bin/pages/chrome       122  language switcher, GitHub mark, route helper
  bin/pages/entities      92  the JSON-LD graph
  bin/pages/locales       32  the languages, and the menu order
  bin/pages/constants     13

Nothing outside had to change. vite.config.ts, six tests and four bin scripts
import `LOCALES`, `MENU_ORDER`, `PAGES`, `generate` and `check`, and the entry
still exports all five -- `design-contract.test.ts` pins that `design-audit.mjs`
imports LOCALES from this path specifically, so the re-export is load-bearing
rather than a courtesy.

The check that this is a pure move: `build-pages.mjs --check` compares a fresh
render against the committed pages, and it passes. The output is byte-identical.

One real regression came with the split and is fixed here. `sitemap-lastmod.mjs`
dates every generated page by the last commit to its markdown source *and to the
generator* -- which it named as `bin/build-pages.mjs`. The templates live in
`bin/pages/` now, so a template change would have stopped dating the pages it
renders, silently: the check would just stop reporting them. It takes the
directory as well now.

Deliberately not re-running `sitemap:lastmod`. It would rewrite 150 lastmod
values to today for pages whose bytes did not change, which is noise in a file
whose whole job is telling crawlers what actually changed.
