+++ bin/pages/render-home.mjs
@@ -1,0 +1,356 @@
+/**
+ * The homepage template.
+ *
+ * Not markdown, because it is not prose: every string sits in a fixed slot and
+ * the decoration around them is layout. content/<locale>/home.json holds the
+ * strings; this holds the structure.
+ */
+import { ORIGIN, REPO } from './constants.mjs';
+import { ID, appEntity, siteEntities, sourceEntity } from './entities.mjs';
+import { DEFAULT_LOCALE, LOCALES } from './locales.mjs';
+import { GH_MARK, langMenu } from './chrome.mjs';
+import { escapeHtml } from './markdown.mjs';
+import { UI } from './ui.mjs';
+
+/**
+ * The homepage.
+ *
+ * It is not prose, so it is not markdown: every string on it sits in a fixed
+ * slot (a hero line, a row of the format index, one of four pillars), and the
+ * decoration around them -- the CSS document window, the reveal classes, the
+ * prefetch hooks -- is layout, not content. content/<locale>/home.json holds
+ * exactly the strings; this holds exactly the structure. Adding a language is
+ * then one JSON file rather than one more 520-line HTML file to keep in step.
+ */
+export function renderHome({ locale, data, locales }) {
+  const L = LOCALES[locale];
+  const ui = UI[locale];
+  const home = L.home;
+  const url = ORIGIN + home;
+  const e = escapeHtml;
+  /**
+   * Links into the app carry the locale. The app's own i18n resolves its
+   * language from `?locale=` first (then cookie, then localStorage, then
+   * navigator.language), so without this a visitor who chose Japanese on the
+   * site would land in an editor guessing from browser settings -- which is
+   * how the hand-written Chinese homepage always did it, and what the first
+   * generated version of this template dropped.
+   */
+  const editor = (query) => `/editor?${locale === DEFAULT_LOCALE ? '' : `locale=${locale}&`}${query}`;
+  const alternates = locales
+    .map((l) => `    <link rel="alternate" hreflang="${l}" href="${ORIGIN + LOCALES[l].home}" />`)
+    .join('\n');
+  const ogAlternates = locales
+    .filter((l) => l !== locale)
+    .map((l) => `    <meta property="og:locale:alternate" content="${LOCALES[l].og}" />`)
+    .join('\n');
+
+  const graph = [
+    ...siteEntities(),
+    appEntity({
+      description: data.description,
+      ...(data.featureList ? { featureList: data.featureList } : {}),
+      ...(data.ecosystem
+        ? { isPartOf: [{ '@id': ID.site }, { '@type': 'SoftwareApplication', ...data.ecosystem }] }
+        : {}),
+    }),
+    // This page: one homepage per language, each pointing at the same app.
+    {
+      '@type': 'WebPage',
+      '@id': `${url}#webpage`,
+      url,
+      name: data.title,
+      description: data.description,
+      inLanguage: L.lang,
+      isPartOf: { '@id': ID.site },
+      about: { '@id': ID.app },
+      primaryImageOfPage: `${ORIGIN}/img/pwa-512.png`,
+    },
+    {
+      '@type': 'FAQPage',
+      '@id': `${url}#faq`,
+      inLanguage: L.lang,
+      mainEntity: data.sections.faq.items.map(({ q, a }) => ({
+        '@type': 'Question',
+        name: q,
+        acceptedAnswer: { '@type': 'Answer', text: a },
+      })),
+    },
+    sourceEntity(),
+  ];
+  const jsonLd = JSON.stringify({ '@context': 'https://schema.org', '@graph': graph }, null, 2)
+    .split('\n')
+    .map((l) => '      ' + l)
+    .join('\n');
+
+  const rows = data.sections.formats.rows
+    .map(
+      (r) => `          <a class="row" href="${e(r.href)}">
+            <span class="ext">${e(r.ext)}</span>
+            <span class="name">${e(r.name)}</span>
+            <span class="desc">${e(r.desc)}</span>
+            <span class="go" aria-hidden="true">&rarr;</span>
+          </a>`,
+    )
+    .join('\n');
+  const pillars = data.sections.pillars.items
+    .map(
+      (p) => `          <div class="pillar">
+            <span class="k">${e(p.k)}</span>
+            <h3>${e(p.h3)}</h3>
+            <p>${e(p.p)}</p>
+          </div>`,
+    )
+    .join('\n');
+  const steps = data.sections.steps.items
+    .map(
+      (st) => `          <div class="step">
+            <div class="n">${e(st.n)}</div>
+            <h3>${e(st.h3)}</h3>
+            <p>${e(st.p)}</p>
+            <span class="loc">${e(st.loc)}</span>
+          </div>`,
+    )
+    .join('\n');
+  const faq = data.sections.faq.items
+    .map(
+      (f, i) => `            <details${i === 0 ? ' open' : ''}>
+              <summary>${e(f.q)}</summary>
+              <p class="a">${e(f.a)}</p>
+            </details>`,
+    )
+    .join('\n');
+  const nav = data.nav.map((l) => `            <a class="navlink" href="${e(l.href)}">${e(l.label)}</a>`).join('\n');
+  const footLinks = data.foot.links.map((l) => `          <a href="${e(l.href)}">${e(l.label)}</a>`).join('\n');
+  const trust = data.trust.map((t) => `<span>${e(t)}</span>`).join('');
+
+  return `<!doctype html>
+<!-- GENERATED by bin/build-pages.mjs from content/${locale}/home.json -- edit the JSON, not this file. -->
+<html lang="${L.lang}" dir="${L.dir}">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
+    <link href="/img/64.png" rel="shortcut icon" />
+    <link rel="icon" type="image/png" href="/img/64.png" />
+    <link rel="manifest" href="/manifest.json" />
+    <!-- Route split: this is a static landing page, /editor hosts the app. Deep
+         links that used to target it (?file= ?src= ?new= ?open=local ?embed=
+         ?embedded= ?readonly ?agent) and any embedding iframe keep working:
+         hand them to /editor with the same query before anything renders. -->
+    <script>
+      (function () {
+        try {
+          var q = location.search;
+          var deep = /[?&](file|src|new|open|embed|embedded|readonly|agent)(=|&|$)/.test(q);
+          var framed = window.parent !== window;
+          if (deep || framed) {
+            location.replace('/editor' + (q || (framed ? '?embed=1' : '')) + location.hash);
+          }
+        } catch (e) {}
+      })();
+    </script>
+    <meta name="theme-color" media="(prefers-color-scheme: light)" content="#ffffff" />
+    <meta name="theme-color" media="(prefers-color-scheme: dark)" content="#000000" />
+    <meta name="mobile-web-app-capable" content="yes" />
+    <meta name="apple-mobile-web-app-status-bar-style" content="black" />
+
+    <title>${e(data.title)}</title>
+    <meta name="description" content="${e(data.description)}" />
+    <link rel="canonical" href="${url}" />
+    <meta name="robots" content="index, follow, max-image-preview:large, max-snippet:-1, max-video-preview:-1" />
+${alternates}
+    <link rel="alternate" hreflang="x-default" href="${ORIGIN + LOCALES[DEFAULT_LOCALE].home}" />
+    <!-- No-flash theme restore: apply a forced light/dark before first paint so a
+         dark-mode visitor never sees a white flash. -->
+    <script>
+      try {
+        var t = localStorage.getItem('ran-theme');
+        if (t === 'dark' || t === 'light') {
+          document.documentElement.setAttribute('data-ran-theme', t);
+          document.documentElement.setAttribute('theme', t);
+        }
+      } catch (e) {}
+    </script>
+    <link rel="stylesheet" href="/ran-fonts/fonts.css" />
+    <link rel="stylesheet" href="/ran-tokens.css" />
+    <link rel="stylesheet" href="/home.css" />
+
+    <meta property="og:type" content="website" />
+    <meta property="og:site_name" content="Online Document Editor" />
+    <meta property="og:locale" content="${L.og}" />
+${ogAlternates}
+    <meta property="og:title" content="${e(data.title)}" />
+    <meta property="og:description" content="${e(data.ogDescription || data.description)}" />
+    <meta property="og:url" content="${url}" />
+    <meta property="og:image" content="${ORIGIN}/img/pwa-512.png" />
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:title" content="${e(data.title)}" />
+    <meta name="twitter:description" content="${e(data.ogDescription || data.description)}" />
+    <meta name="twitter:image" content="${ORIGIN}/img/pwa-512.png" />
+
+    <script type="application/ld+json">
+${jsonLd}
+    </script>
+
+    <script src="/ranui-iife/button.iife.js" defer></script>
+    <script src="/ranui-iife/popover.iife.js" defer></script>
+    <script src="/ranui-iife/content.iife.js" defer></script>
+    <script src="/ranui-iife/theme-switch.iife.js" defer></script>
+    <script src="/lang-switch.js" defer></script>
+    <script src="/open-local.js" defer></script>
+    <script src="/landing-prefetch.js" defer></script>
+    <script src="/history-recent.js" defer></script>
+  </head>
+
+  <body>
+    <svg width="0" height="0" style="position: absolute" aria-hidden="true">
+      <symbol id="gh-mark" viewBox="0 0 16 16">
+        <path d="${GH_MARK}" />
+      </symbol>
+    </svg>
+    <section id="landing-hero">
+      <header class="bar">
+        <div class="wrap">
+          <a class="brand" href="${home}"><span class="logo">D</span>${e(ui.siteName)}</a>
+          <nav>
+${nav}
+            <a class="navlink gh" href="${REPO}" rel="noopener" target="_blank">
+              <svg class="ghmark" aria-hidden="true"><use href="#gh-mark"></use></svg> GitHub
+            </a>
+${langMenu(locale, locales, ui, (l) => LOCALES[l].home)}
+          </nav>
+        </div>
+      </header>
+
+      <div class="hero wrap">
+        <div class="hero-copy">
+          <span class="chip reveal d1"><span class="dot"></span>${data.chip}</span>
+          <h1 class="reveal d2">${e(data.h1.plain)} <span class="accent">${e(data.h1.accent)}</span></h1>
+          <p class="sub reveal d3">${e(data.sub)}</p>
+          <div class="cta reveal d4">
+            <r-button type="primary" id="hero-open" data-open-local="${editor('open=local')}">${e(data.cta.open)}</r-button>
+            <a href="${editor('new=docx')}" data-prefetch="docx"><r-button id="hero-new-docx">${e(data.cta.docx)}</r-button></a>
+            <a href="${editor('new=xlsx')}" data-prefetch="xlsx"><r-button id="hero-new-xlsx">${e(data.cta.xlsx)}</r-button></a>
+            <a href="${editor('new=pptx')}" data-prefetch="pptx"><r-button id="hero-new-pptx">${e(data.cta.pptx)}</r-button></a>
+          </div>
+          <!-- Autosave is a promise about the user's data, so it is served HTML
+               rather than drawn by script: the retention window and the way to
+               inspect and delete what is kept are visible to a first-time
+               visitor, to a crawler, and with JavaScript off. -->
+          <div class="recent reveal d5">
+            <span data-recent-slot data-recent-label="${e(data.recent.label)}" hidden></span>
+            <span class="recent-note">${e(data.recent.note)}</span>
+            <a class="recent-all" href="${locale === DEFAULT_LOCALE ? '/history' : `/history?locale=${locale}`}"
+              >${e(data.recent.all)}</a
+            >
+          </div>
+          <div class="trust reveal d5">${trust}</div>
+        </div>
+
+        <!-- Decorative "document being edited locally" window -- pure CSS, no JS. -->
+        <div class="docwin reveal d3" aria-hidden="true">
+          <div class="dw-bar">
+            <span class="dw-dots"><i></i><i></i><i></i></span>
+            <span class="dw-file">${e(data.docwin.file)}</span>
+            <span class="dw-badge">${e(data.docwin.badge)}</span>
+          </div>
+          <div class="dw-tabs"><span class="on">.docx</span><span>.xlsx</span><span>.pptx</span><span>.csv</span></div>
+          <div class="dw-body">
+            <div class="dw-sheet">
+              <span class="dl t"></span>
+              <span class="dl w92"></span>
+              <span class="dl w78"></span>
+              <p class="dw-text"><mark>${e(data.docwin.selected)}</mark><i class="caret"></i></p>
+              <span class="dl w85"></span>
+              <span class="dl w60"></span>
+              <span class="dl t2"></span>
+              <span class="dl w88"></span>
+              <span class="dl w92"></span>
+              <span class="dl w70"></span>
+            </div>
+            <aside class="dw-note">
+              <b>${e(data.docwin.noteAuthor)}</b>
+              <span>${e(data.docwin.noteBody)}</span>
+            </aside>
+          </div>
+          <div class="dw-net"><i class="ok"></i>${e(data.docwin.net)}</div>
+        </div>
+      </div>
+
+      <div class="section wrap">
+        <div class="section-head">
+          <span class="eyebrow">${e(data.sections.formats.eyebrow)}</span>
+          <h2>${e(data.sections.formats.h2)}</h2>
+        </div>
+        <div class="index">
+${rows}
+        </div>
+      </div>
+
+      <div class="section wrap">
+        <div class="section-head">
+          <span class="eyebrow">${e(data.sections.pillars.eyebrow)}</span>
+          <h2>${e(data.sections.pillars.h2)}</h2>
+        </div>
+        <div class="pillars">
+${pillars}
+        </div>
+      </div>
+
+      <div class="section wrap">
+        <div class="section-head">
+          <span class="eyebrow">${e(data.sections.steps.eyebrow)}</span>
+          <h2>${e(data.sections.steps.h2)}</h2>
+        </div>
+        <div class="steps">
+${steps}
+        </div>
+      </div>
+
+      <div class="section wrap faq-sec">
+        <div class="section-head">
+          <span class="eyebrow">${e(data.sections.faq.eyebrow)}</span>
+          <h2>${e(data.sections.faq.h2)}</h2>
+          <p class="head-note">${data.sections.faq.note}</p>
+        </div>
+        <div class="faq">
+          <div class="qa">
+${faq}
+          </div>
+        </div>
+      </div>
+
+      <div class="eco">
+        <div class="wrap">
+          <span class="eco-label">
+            <svg class="gitmark" aria-hidden="true"><use href="#gh-mark"></use></svg>
+            ${e(data.eco.label)}
+          </span>
+          <nav>
+            <span class="here"><b>${e(data.eco.here.name)}</b> <small>${e(data.eco.here.role)}</small></span>
+            <a href="${e(data.eco.other.href)}" rel="noopener" target="_blank">
+              <b>${e(data.eco.other.name)}</b> <small>${e(data.eco.other.role)}</small>
+            </a>
+          </nav>
+        </div>
+      </div>
+
+      <footer class="foot wrap">
+        <span>${e(data.foot.copy)}</span>
+        <nav>
+${footLinks}
+        </nav>
+        <r-theme-switch class="theme-switch" label="${e(ui.themeLabel)}"></r-theme-switch>
+        <span class="lic">${e(data.foot.license)}</span>
+        <p class="tm">${e(ui.trademark)}</p>
+      </footer>
+    </section>
+
+    <!-- Registers the service worker (so the landing works offline / installs as
+         an app) and promotes a waiting one, which is what makes a deploy reach a
+         returning visitor. See public/sw-register.js. -->
+    <script src="/sw-register.js" defer></script>
+  </body>
+</html>
+`;
+}
