+++ bin/pages/ui.mjs
@@ -1,0 +1,227 @@
+/**
+ * Per-locale chrome strings: everything in the shell that is words rather
+ * than content. The pages themselves come from markdown; this is the frame
+ * around them.
+ */
+import { REPO } from './constants.mjs';
+
+export const UI = {
+  en: {
+    siteName: 'Document Editor',
+    langAria: 'Language',
+    themeLabel: 'Theme',
+    home: 'Home',
+    openEditor: 'Open editor',
+    onThisPage: 'On this page',
+    more: 'Related pages',
+    sideNote: 'Runs on your device. No upload, no sign-up.',
+    footer: [
+      ['/', 'Open editor'],
+      ['/help', 'Help'],
+      ['/changelog', 'Changelog'],
+      ['/open/docx', 'Open DOCX'],
+      ['/open/xlsx', 'Open XLSX'],
+      ['/open/pptx', 'Open PPTX'],
+      ['/open/pdf', 'Open PDF'],
+      ['/convert/xlsx-to-csv', 'XLSX to CSV'],
+      ['/convert/csv-to-xlsx', 'CSV to XLSX'],
+      ['/no-signup-document-editor', 'No sign-up'],
+      ['/private-document-editor', 'Private editor'],
+      ['/offline-document-editor', 'Offline'],
+      ['/onlyoffice-online-free', 'ONLYOFFICE online'],
+      ['/edit-documents-without-account', 'No account'],
+      ['/embed-document-editor', 'Embed API'],
+    ],
+    generatedNote: (src) => `Source: ${src} in the repository`,
+    ossNote: `<strong>Open source &amp; self-hostable.</strong> Under AGPL-3.0 — verify that nothing is uploaded, or run your own copy: <a href="${REPO}" rel="noopener">github.com/ranuts/document</a>.`,
+    trademark: `ONLYOFFICE is a trademark of Ascensio System SIA. This site is not an official ONLYOFFICE product and is not affiliated with or endorsed by Ascensio System SIA.`,
+  },
+  'zh-CN': {
+    siteName: 'Document Editor',
+    langAria: '语言',
+    themeLabel: '主题',
+    home: '首页',
+    openEditor: '打开编辑器',
+    onThisPage: '本页目录',
+    more: '相关页面',
+    sideNote: '在你的设备上运行，不上传、免注册。',
+    footer: [
+      ['/zh-CN/', '打开编辑器'],
+      ['/zh-CN/help', '帮助'],
+      ['/zh-CN/changelog', '更新日志'],
+      ['/zh-CN/open/docx', '打开 DOCX'],
+      ['/zh-CN/open/xlsx', '打开 XLSX'],
+      ['/zh-CN/open/pptx', '打开 PPTX'],
+      ['/zh-CN/open/pdf', '打开 PDF'],
+      ['/zh-CN/convert/xlsx-to-csv', 'XLSX 转 CSV'],
+      ['/zh-CN/convert/csv-to-xlsx', 'CSV 转 XLSX'],
+      ['/zh-CN/no-signup-document-editor', '免注册'],
+      ['/zh-CN/private-document-editor', '隐私编辑器'],
+      ['/zh-CN/offline-document-editor', '离线'],
+      ['/zh-CN/onlyoffice-online-free', 'OnlyOffice 在线版'],
+      ['/zh-CN/edit-documents-without-account', '免账号'],
+      ['/zh-CN/embed-document-editor', 'Embed API'],
+    ],
+    generatedNote: (src) => `来源：仓库中的 ${src}`,
+    ossNote: `<strong>开源 · 可自托管。</strong>采用 AGPL-3.0——你可以核实没有任何上传，或者自建一份：<a href="${REPO}" rel="noopener">github.com/ranuts/document</a>。`,
+    trademark: `ONLYOFFICE 是 Ascensio System SIA 的商标。本站并非官方 ONLYOFFICE 产品，与 Ascensio System SIA 无隶属关系，也未获其背书。`,
+  },
+  ja: {
+    siteName: 'Document Editor',
+    langAria: '言語',
+    themeLabel: 'テーマ',
+    home: 'ホーム',
+    openEditor: 'エディタを開く',
+    onThisPage: 'このページの目次',
+    more: '関連ページ',
+    sideNote: 'お使いの端末で動作します。アップロードなし、登録不要。',
+    footer: [
+      ['/ja/', 'エディタを開く'],
+      ['/ja/help', 'ヘルプ'],
+      ['/ja/changelog', '変更履歴'],
+      ['/ja/open/docx', 'DOCX を開く'],
+      ['/ja/open/xlsx', 'XLSX を開く'],
+      ['/ja/open/pptx', 'PPTX を開く'],
+      ['/ja/open/pdf', 'PDF を開く'],
+      ['/ja/convert/xlsx-to-csv', 'XLSX を CSV に'],
+      ['/ja/convert/csv-to-xlsx', 'CSV を XLSX に'],
+      ['/ja/no-signup-document-editor', '登録不要'],
+      ['/ja/private-document-editor', 'プライバシー'],
+      ['/ja/offline-document-editor', 'オフライン'],
+      ['/ja/edit-documents-without-account', 'アカウント不要'],
+      ['/ja/embed-document-editor', 'Embed API'],
+    ],
+    generatedNote: (src) => `ソース: リポジトリの ${src}`,
+    ossNote: `<strong>オープンソース・セルフホスト可能。</strong>AGPL-3.0 のもとで公開——何もアップロードされないことを自分で確認でき、自分で運用することもできます: <a href="${REPO}" rel="noopener">github.com/ranuts/document</a>。`,
+    trademark: `ONLYOFFICE は Ascensio System SIA の商標です。本サイトは公式の ONLYOFFICE 製品ではなく、Ascensio System SIA との提携も推奨関係もありません。`,
+  },
+  de: {
+    siteName: 'Document Editor',
+    langAria: 'Sprache',
+    themeLabel: 'Design',
+    home: 'Startseite',
+    openEditor: 'Editor öffnen',
+    onThisPage: 'Auf dieser Seite',
+    more: 'Verwandte Seiten',
+    sideNote: 'Läuft auf Ihrem Gerät. Kein Upload, keine Anmeldung.',
+    footer: [
+      ['/de/', 'Editor öffnen'],
+      ['/de/help', 'Hilfe'],
+      ['/de/changelog', 'Änderungen'],
+      ['/de/open/docx', 'DOCX öffnen'],
+      ['/de/open/xlsx', 'XLSX öffnen'],
+      ['/de/open/pptx', 'PPTX öffnen'],
+      ['/de/open/pdf', 'PDF öffnen'],
+      ['/de/convert/xlsx-to-csv', 'XLSX zu CSV'],
+      ['/de/convert/csv-to-xlsx', 'CSV zu XLSX'],
+      ['/de/no-signup-document-editor', 'Ohne Anmeldung'],
+      ['/de/private-document-editor', 'Datenschutz'],
+      ['/de/offline-document-editor', 'Offline'],
+      ['/de/edit-documents-without-account', 'Ohne Konto'],
+      ['/de/embed-document-editor', 'Embed API'],
+    ],
+    generatedNote: (src) => `Quelle: ${src} im Repository`,
+    ossNote: `<strong>Open Source &amp; selbst hostbar.</strong> Unter AGPL-3.0 — prüfen Sie selbst, dass nichts hochgeladen wird, oder betreiben Sie eine eigene Kopie: <a href="${REPO}" rel="noopener">github.com/ranuts/document</a>.`,
+    trademark: `ONLYOFFICE ist eine Marke von Ascensio System SIA. Diese Website ist kein offizielles ONLYOFFICE-Produkt und steht in keiner Verbindung zu Ascensio System SIA.`,
+  },
+  es: {
+    siteName: 'Document Editor',
+    langAria: 'Idioma',
+    themeLabel: 'Tema',
+    home: 'Inicio',
+    openEditor: 'Abrir el editor',
+    onThisPage: 'En esta página',
+    more: 'Páginas relacionadas',
+    sideNote: 'Funciona en tu dispositivo. Sin subidas, sin registro.',
+    footer: [
+      ['/es/', 'Abrir el editor'],
+      ['/es/help', 'Ayuda'],
+      ['/es/changelog', 'Novedades'],
+      ['/es/open/docx', 'Abrir DOCX'],
+      ['/es/open/xlsx', 'Abrir XLSX'],
+      ['/es/open/pptx', 'Abrir PPTX'],
+      ['/es/open/pdf', 'Abrir PDF'],
+      ['/es/convert/xlsx-to-csv', 'XLSX a CSV'],
+      ['/es/convert/csv-to-xlsx', 'CSV a XLSX'],
+      ['/es/no-signup-document-editor', 'Sin registro'],
+      ['/es/private-document-editor', 'Privacidad'],
+      ['/es/offline-document-editor', 'Sin conexión'],
+      ['/es/edit-documents-without-account', 'Sin cuenta'],
+      ['/es/embed-document-editor', 'Embed API'],
+    ],
+    generatedNote: (src) => `Fuente: ${src} en el repositorio`,
+    ossNote: `<strong>Código abierto y autoalojable.</strong> Bajo AGPL-3.0: comprueba que no se sube nada, o ejecuta tu propia copia: <a href="${REPO}" rel="noopener">github.com/ranuts/document</a>.`,
+    trademark: `ONLYOFFICE es una marca de Ascensio System SIA. Este sitio no es un producto oficial de ONLYOFFICE ni está afiliado a Ascensio System SIA ni respaldado por ella.`,
+  },
+  ko: {
+    siteName: 'Document Editor',
+    langAria: '언어',
+    themeLabel: '테마',
+    home: '홈',
+    openEditor: '편집기 열기',
+    onThisPage: '이 페이지의 목차',
+    more: '관련 페이지',
+    sideNote: '기기에서 실행됩니다. 업로드 없음, 가입 불필요.',
+    footer: [
+      ['/ko/', '편집기 열기'],
+      ['/ko/help', '도움말'],
+      ['/ko/changelog', '변경 내역'],
+      ['/ko/open/docx', 'DOCX 열기'],
+      ['/ko/open/xlsx', 'XLSX 열기'],
+      ['/ko/open/pptx', 'PPTX 열기'],
+      ['/ko/open/pdf', 'PDF 열기'],
+      ['/ko/convert/xlsx-to-csv', 'XLSX를 CSV로'],
+      ['/ko/convert/csv-to-xlsx', 'CSV를 XLSX로'],
+      ['/ko/no-signup-document-editor', '가입 불필요'],
+      ['/ko/private-document-editor', '개인정보'],
+      ['/ko/offline-document-editor', '오프라인'],
+      ['/ko/edit-documents-without-account', '계정 불필요'],
+      ['/ko/embed-document-editor', 'Embed API'],
+    ],
+    generatedNote: (src) => `출처: 저장소의 ${src}`,
+    ossNote: `<strong>오픈 소스이며 직접 호스팅할 수 있습니다.</strong> AGPL-3.0으로 공개되어 있어, 아무것도 업로드되지 않는다는 것을 직접 확인하거나 직접 운영할 수 있습니다: <a href="${REPO}" rel="noopener">github.com/ranuts/document</a>.`,
+    trademark: `ONLYOFFICE는 Ascensio System SIA의 상표입니다. 이 사이트는 공식 ONLYOFFICE 제품이 아니며 Ascensio System SIA와 제휴하거나 후원받지 않았습니다.`,
+  },
+  pt: {
+    siteName: 'Document Editor',
+    langAria: 'Idioma',
+    themeLabel: 'Tema',
+    home: 'Início',
+    openEditor: 'Abrir o editor',
+    onThisPage: 'Nesta página',
+    more: 'Páginas relacionadas',
+    sideNote: 'Roda no seu dispositivo. Sem uploads, sem cadastro.',
+    footer: [
+      ['/pt/', 'Abrir o editor'],
+      ['/pt/help', 'Ajuda'],
+      ['/pt/changelog', 'Novidades'],
+      ['/pt/open/docx', 'Abrir DOCX'],
+      ['/pt/open/xlsx', 'Abrir XLSX'],
+      ['/pt/open/pptx', 'Abrir PPTX'],
+      ['/pt/open/pdf', 'Abrir PDF'],
+      ['/pt/convert/xlsx-to-csv', 'XLSX para CSV'],
+      ['/pt/convert/csv-to-xlsx', 'CSV para XLSX'],
+      ['/pt/no-signup-document-editor', 'Sem cadastro'],
+      ['/pt/private-document-editor', 'Privacidade'],
+      ['/pt/offline-document-editor', 'Offline'],
+      ['/pt/edit-documents-without-account', 'Sem conta'],
+      ['/pt/embed-document-editor', 'Embed API'],
+    ],
+    generatedNote: (src) => `Fonte: ${src} no repositório`,
+    ossNote: `<strong>Código aberto e auto-hospedável.</strong> Sob a AGPL-3.0 — confira você mesmo que nada é enviado, ou rode a sua própria cópia: <a href="${REPO}" rel="noopener">github.com/ranuts/document</a>.`,
+    trademark: `ONLYOFFICE é uma marca da Ascensio System SIA. Este site não é um produto oficial do ONLYOFFICE nem tem afiliação ou endosso da Ascensio System SIA.`,
+  },
+};
+
+/**
+ * Pages to generate. `sources` may point at any markdown file in the repo;
+ * frontmatter (title / description / eyebrow / breadcrumb) is optional and
+ * can be overridden per locale in `meta`. When a locale has no source of its
+ * own it may reuse another (e.g. the changelog is maintained in English only)
+ * -- `notice` is then rendered above the body.
+ */
+/**
+ * Landing-page slugs, in the order they were written. The list is explicit
+ * rather than a directory scan so a stray markdown file cannot quietly become
+ * a public page, and so the order the sitemap and llms.txt see is stable.
+ */
